+++ CHANGES.md
@@ -4,6 +4,12 @@
 
 None yet.
 
+## v1.1.0 (2015-09-02)
+
+* #6 a new suite of hrtime manipulation routines: `hrtimeAdd()`,
+  `hrtimeAccum()`, `hrtimeNanosec()`, `hrtimeMicrosec()` and
+  `hrtimeMillisec()`.
+
 ## v1.0.0 (2015-09-01)
 
 First tracked release.  Includes everything in previous releases, plus: